requirements.txt
pygame
python-socketio
python-socketio[client]